backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
python-multipart==0.0.6
python-dotenv==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.11.5
pydantic-settings==2.1.0
cloudinary==1.37.0
llama-parse==0.6.54
google-generativeai==0.8.3
chromadb==0.5.20
aiofiles==23.2.0
asyncpg==0.29.0
llama-index-core==0.12.0
llama-index-vector-stores-chroma==0.3.3
llama-index-llms-google-genai==0.3.0
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2
httpx==0.28.1
anyio==4.8.0
